% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/githubCalendar.R
\name{githubCalendar-shiny}
\alias{githubCalendar-shiny}
\alias{githubCalendar}
\alias{githubCalendarOutput}
\alias{renderGithubCalendar}
\title{A github inspired calendar visualization}
\usage{
githubCalendar(dates, values, width = "100\%", height = NULL, elementId = NULL)

githubCalendarOutput(outputId, width = "100\%", height = "400px")

renderGithubCalendar(dates, values, expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{dates}{the dates to plot on the calendar, must all be in the same year}

\item{values}{the values associated with each date}

\item{width, height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{elementId}{output variable to read from}

\item{outputId}{output variable to read from}

\item{expr}{An expression that generates a githubCalendar}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}
}
\description{
Output and render functions for using githubCalendar within Shiny
applications and interactive Rmd documents.
}
