% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate_dates.R
\name{interpolate_dates}
\alias{interpolate_dates}
\title{Interpolate Missing Dates}
\usage{
interpolate_dates(dates, values)
}
\arguments{
\item{dates}{a vector of dates, must be only one year}

\item{values}{a vector of values for dates}
}
\value{
a dataframe with dates and values
}
\description{
Create a table of dates and associated values
then add missing dates to table with a value of 0
}
